finance calculator file corrected

+++ python_introduction/finance_calculator.py
@@ -1,5 +1,5 @@
-monthly_input = int(input("Enter your monthly income: "))
-monthly_expenses = int(input("Enter your total monthly expenses: "))
+monthly_input = float(input("Enter your monthly income: "))
+monthly_expenses = float(input("Enter your total monthly expenses: "))
 monthly_savings = monthly_input - monthly_expenses
 annual_interest_rate = 0.05 
 projected_savings = monthly_savings * 12 + ( monthly_savings * 12 * annual_interest_rate)
